Feat: 코드에 API 변경 사항 반영 (#110)
* Feat: ì관리자 상품 조ì회 시 초기 재고와 판매된 재고 모두 조회

* Feat: ê°dto 가격을 ㅑint -> bigdecimal 변경

* Chore: 엔드포포인ㅇ트 볈형으로 변경

+++ src/main/java/com/jelly/zzirit/domain/adminItem/dto/request/ItemCreateRequest.java
@@ -11,7 +11,7 @@ import java.math.BigDecimal;
 public record ItemCreateRequest (
     @NotBlank String name, // validity check
     @PositiveOrZero int stockQuantity,
-    @PositiveOrZero int price,
+    @PositiveOrZero BigDecimal price,
     @NotNull Long typeId,
     @NotNull Long brandId,
     @NotBlank String imageUrl
@@ -20,7 +20,7 @@ public record ItemCreateRequest (
         return Item.builder()
                 .name(name)
                 .imageUrl(imageUrl)
-                .price(BigDecimal.valueOf(price)) // todo: int->bigdecimal로 변경 필요
+                .price(price) // todo: int->bigdecimal로 변경 필요
                 .typeBrand(typeBrand)
                 .itemStatus(ItemStatus.NONE) // todo: item status 업데이트 로직 추가
                 .build();
